requirements.txt
chainlit==2.4.400
langchain==0.3.23
langchain_cohere==0.4.3
langchain_community==0.3.21
langchain_core==0.3.51
python-dotenv==1.1.0
